+++ component/notify/config_server_test.go
@@ -88,15 +88,6 @@ func runErrorResponse() *httptest.Server {
 	return ts
 }
 
-func runLongNotmodifiedConfigResponse() *httptest.Server {
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		time.Sleep(500 * time.Microsecond)
-		w.WriteHeader(http.StatusNotModified)
-	}))
-
-	return ts
-}
-
 //Error response
 //will hold 5s and keep response 404
 func runErrorConfigResponse() *httptest.Server {
