+++ repository_test.go
@@ -1,0 +1,196 @@
+package agollo
+
+import (
+	"fmt"
+	. "github.com/tevid/gohamcrest"
+	"github.com/zouyx/agollo/v2/component/notify"
+	"github.com/zouyx/agollo/v2/env"
+	"github.com/zouyx/agollo/v2/storage"
+	"net/http"
+	"net/http/httptest"
+	"testing"
+	"time"
+)
+
+//init param
+func init() {
+}
+
+
+func createMockApolloConfig(expireTime int) map[string]string {
+	configs := make(map[string]string, 0)
+	//string
+	configs["string"] = "value"
+	//int
+	configs["int"] = "1"
+	//float
+	configs["float"] = "190.3"
+	//bool
+	configs["bool"] = "true"
+
+	storage.UpdateApolloConfigCache(configs, expireTime, storage.GetDefaultNamespace())
+
+	return configs
+}
+
+
+func TestGetConfigValueNullApolloConfig(t *testing.T) {
+	//clear Configurations
+	defaultConfigCache := GetDefaultConfigCache()
+	defaultConfigCache.Clear()
+
+	//test getValue
+	value := GetValue("joe")
+
+	Assert(t, "", Equal(value))
+
+	//test GetStringValue
+	defaultValue := "j"
+
+	//test default
+	v := GetStringValue("joe", defaultValue)
+
+	Assert(t, defaultValue, Equal(v))
+
+}
+
+func TestGetIntValue(t *testing.T) {
+	createMockApolloConfig(120)
+	defaultValue := 100000
+
+	//test default
+	v := GetIntValue("joe", defaultValue)
+
+	Assert(t, defaultValue, Equal(v))
+
+	//normal value
+	v = GetIntValue("int", defaultValue)
+
+	Assert(t, 1, Equal(v))
+
+	//error type
+	v = GetIntValue("float", defaultValue)
+
+	Assert(t, defaultValue, Equal(v))
+}
+
+func TestGetFloatValue(t *testing.T) {
+	defaultValue := 100000.1
+
+	//test default
+	v := GetFloatValue("joe", defaultValue)
+
+	Assert(t, defaultValue, Equal(v))
+
+	//normal value
+	v = GetFloatValue("float", defaultValue)
+
+	Assert(t, 190.3, Equal(v))
+
+	//error type
+	v = GetFloatValue("int", defaultValue)
+
+	Assert(t, float64(1), Equal(v))
+}
+
+func TestGetBoolValue(t *testing.T) {
+	defaultValue := false
+
+	//test default
+	v := GetBoolValue("joe", defaultValue)
+
+	Assert(t, defaultValue, Equal(v))
+
+	//normal value
+	v = GetBoolValue("bool", defaultValue)
+
+	Assert(t, true, Equal(v))
+
+	//error type
+	v = GetBoolValue("float", defaultValue)
+
+	Assert(t, defaultValue, Equal(v))
+}
+
+func TestGetStringValue(t *testing.T) {
+	defaultValue := "j"
+
+	//test default
+	v := GetStringValue("joe", defaultValue)
+
+	Assert(t, defaultValue, Equal(v))
+
+	//normal value
+	v = GetStringValue("string", defaultValue)
+
+	Assert(t, "value", Equal(v))
+}
+
+
+func TestAutoSyncConfigServicesNormal2NotModified(t *testing.T) {
+	server := runLongNotmodifiedConfigResponse()
+	newAppConfig := getTestAppConfig()
+	newAppConfig.Ip = server.URL
+	time.Sleep(1 * time.Second)
+	appConfig := env.GetPlainAppConfig()
+	appConfig.NextTryConnTime = 0
+
+	notify.AutoSyncConfigServicesSuccessCallBack([]byte(configResponseStr))
+
+	config := env.GetCurrentApolloConfig()[newAppConfig.NamespaceName]
+
+	fmt.Println("sleeping 10s")
+
+	time.Sleep(10 * time.Second)
+
+	fmt.Println("checking agcache time left")
+	defaultConfigCache := GetDefaultConfigCache()
+
+	defaultConfigCache.Range(func(key, value interface{}) bool {
+		Assert(t, string(value.([]byte)), NotNilVal())
+		return true
+	})
+
+	Assert(t, "100004458", Equal(config.AppId))
+	Assert(t, "default", Equal(config.Cluster))
+	Assert(t, "application", Equal(config.NamespaceName))
+	Assert(t, "20170430092936-dee2d58e74515ff3", Equal(config.ReleaseKey))
+	Assert(t, "value1", Equal(GetStringValue("key1", "")))
+	Assert(t, "value2", Equal(GetStringValue("key2", "")))
+
+	err := notify.AutoSyncConfigServices(newAppConfig)
+
+	fmt.Println("checking agcache time left")
+	defaultConfigCache.Range(func(key, value interface{}) bool {
+		Assert(t, string(value.([]byte)), NotNilVal())
+		return true
+	})
+
+	fmt.Println(err)
+
+	//sleep for async
+	time.Sleep(1 * time.Second)
+	checkBackupFile(t)
+}
+
+
+func checkBackupFile(t *testing.T) {
+	appConfig := env.GetPlainAppConfig()
+	newConfig, e := env.LoadConfigFile(appConfig.GetBackupConfigPath(), "application")
+	t.Log(newConfig.Configurations)
+	Assert(t, e, NilVal())
+	Assert(t, newConfig.Configurations, NotNilVal())
+	for k, v := range newConfig.Configurations {
+		Assert(t, GetStringValue(k, ""), Equal(v))
+	}
+}
+
+
+func runLongNotmodifiedConfigResponse() *httptest.Server {
+	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		time.Sleep(500 * time.Microsecond)
+		w.WriteHeader(http.StatusNotModified)
+	}))
+
+	return ts
+}
