+++ storage/repository_test.go
@@ -24,7 +24,7 @@ func createMockApolloConfig(expireTime int) map[string]string {
 	//bool
 	configs["bool"] = "true"
 
-	updateApolloConfigCache(configs, expireTime, defaultNamespace)
+	UpdateApolloConfigCache(configs, expireTime, defaultNamespace)
 
 	return configs
 }
@@ -53,98 +53,6 @@ func TestUpdateApolloConfigNull(t *testing.T) {
 func TestGetApolloConfigCache(t *testing.T) {
 	cache := GetApolloConfigCache()
 	Assert(t, cache, NotNilVal())
-}
-
-func TestGetConfigValueNullApolloConfig(t *testing.T) {
-	//clear Configurations
-	defaultConfigCache := GetDefaultConfigCache()
-	defaultConfigCache.Clear()
-
-	//test getValue
-	value := GetValue("joe")
-
-	Assert(t, "", Equal(value))
-
-	//test GetStringValue
-	defaultValue := "j"
-
-	//test default
-	v := GetStringValue("joe", defaultValue)
-
-	Assert(t, defaultValue, Equal(v))
-
-}
-
-func TestGetIntValue(t *testing.T) {
-	createMockApolloConfig(configCacheExpireTime)
-	defaultValue := 100000
-
-	//test default
-	v := GetIntValue("joe", defaultValue)
-
-	Assert(t, defaultValue, Equal(v))
-
-	//normal value
-	v = GetIntValue("int", defaultValue)
-
-	Assert(t, 1, Equal(v))
-
-	//error type
-	v = GetIntValue("float", defaultValue)
-
-	Assert(t, defaultValue, Equal(v))
-}
-
-func TestGetFloatValue(t *testing.T) {
-	defaultValue := 100000.1
-
-	//test default
-	v := GetFloatValue("joe", defaultValue)
-
-	Assert(t, defaultValue, Equal(v))
-
-	//normal value
-	v = GetFloatValue("float", defaultValue)
-
-	Assert(t, 190.3, Equal(v))
-
-	//error type
-	v = GetFloatValue("int", defaultValue)
-
-	Assert(t, float64(1), Equal(v))
-}
-
-func TestGetBoolValue(t *testing.T) {
-	defaultValue := false
-
-	//test default
-	v := GetBoolValue("joe", defaultValue)
-
-	Assert(t, defaultValue, Equal(v))
-
-	//normal value
-	v = GetBoolValue("bool", defaultValue)
-
-	Assert(t, true, Equal(v))
-
-	//error type
-	v = GetBoolValue("float", defaultValue)
-
-	Assert(t, defaultValue, Equal(v))
-}
-
-func TestGetStringValue(t *testing.T) {
-	defaultValue := "j"
-
-	//test default
-	v := GetStringValue("joe", defaultValue)
-
-	Assert(t, defaultValue, Equal(v))
-
-	//normal value
-	v = GetStringValue("string", defaultValue)
-
-	Assert(t, "value", Equal(v))
 }
 
 func TestConfig_GetStringValue(t *testing.T) {
